DB통합
