+++ lib/Sin/WPage.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:sqflite/sqflite.dart';
 import 'package:path/path.dart';
-import 'Commu.dart';
+import '../health.dart';
 
 class WritePage extends StatefulWidget {
   @override
