+++ lib/Sin/community.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:health/Sin/Commu.dart';
+import '../health.dart';
 import 'posts_list.dart';
 import 'search.dart'; // CustomSearchDelegate import
