+++ lib/health.dart
@@ -1,7 +1,40 @@
-import 'package:flutter/material.dart';
 import 'package:sqflite/sqflite.dart';
 import 'package:path/path.dart';
 
+// Member 클래스 정의
+class Member {
+  final int memberNumber; //회원번호
+  final String password; //비밀번호
+  final String name; //이름
+  final String phoneNumber; //전화번호
+  final DateTime registrationDate; //등록일
+  final DateTime expirationDate; //마감일
+  final String memberState; //회원권상태: 정상/정지
+
+  Member({
+    required this.memberNumber,
+    required this.password,
+    required this.name,
+    required this.phoneNumber,
+    required this.registrationDate,
+    required this.expirationDate,
+    required this.memberState,
+  });
+
+  Map<String, dynamic> toMap() {
+    return {
+      'memberNumber': memberNumber,
+      'password': password,
+      'name': name,
+      'phoneNumber': phoneNumber,
+      'registrationDate': registrationDate.toIso8601String(),
+      'expirationDate': expirationDate.toIso8601String(),
+      'memberState': memberState,
+    };
+  }
+}
+
+// Commu 클래스 정의
 class Commu {
   final int? postID;
   final int? fk_memberNumber;
@@ -29,7 +62,6 @@ class Commu {
     this.name,
   });
 
-  // 맵 형식으로 변환
   Map<String, dynamic> toMap() {
     return {
       'postID': postID,
@@ -45,7 +77,7 @@ class Commu {
   }
 }
 
-// 댓글 클래스 Comment 정의
+// Comment 클래스 정의
 class Comment {
   final int? commentID;
   final int postID;
@@ -61,7 +93,6 @@ class Comment {
     required this.createdAt,
   });
 
-  // 맵 형식으로 변환
   Map<String, dynamic> toMap() {
     return {
       'commentID': commentID,
@@ -73,37 +104,45 @@ class Comment {
   }
 }
 
-// DatabaseHelper 클래스 정의
+// 통합된 데이터베이스 헬퍼 클래스 정의
 class DatabaseHelper {
-  static late Database _database;
-  static const String commuDbName = 'commu.db';
+  static Database? _database;
+  static const String dbName = 'health.db';
+  static const String membersTable = 'members';
   static const String postsTable = 'Posts';
   static const String commentsTable = 'Comments';
 
-  // 데이터베이스 초기화
-  static Future<Database> initDatabase() async {
-    _database = await openDatabase(
-      join(await getDatabasesPath(), commuDbName),
-      onCreate: (db, version) async {
-        await _createTables(db);
-        await _insertInitialData(db);
-      },
-      version: 3, // 버전 변경
-      onUpgrade: (db, oldVersion, newVersion) async {
-        if (oldVersion < 2) {
-          await _createTables(db);
-        }
-        if (oldVersion < 3) {
-          await db.execute(
-              'ALTER TABLE $postsTable ADD COLUMN reportCount INTEGER DEFAULT 0'); // 신고 수 필드 추가
-        }
-      },
-    );
-    return _database;
+  static Future<Database> get database async {
+    if (_database != null) {
+      return _database!;
+    }
+    _database = await initDatabase();
+    return _database!;
   }
 
-  // 테이블 생성
+  static Future<Database> initDatabase() async {
+    final databasesPath = await getDatabasesPath();
+    final path = join(databasesPath, dbName);
+    print('Database path: $path');
+    return openDatabase(path, version: 1, onCreate: (db, version) async {
+      await _createTables(db);
+      await _insertInitialData(db);
+    });
+  }
+
   static Future<void> _createTables(Database db) async {
+    await db.execute('''
+      CREATE TABLE IF NOT EXISTS $membersTable(
+        memberNumber INTEGER PRIMARY KEY AUTOINCREMENT,
+        password TEXT,
+        name TEXT,
+        phoneNumber TEXT,
+        registrationDate TEXT,
+        expirationDate TEXT,
+        memberState TEXT
+      )
+    ''');
+
     await db.execute('''
       CREATE TABLE IF NOT EXISTS $postsTable(
         postID INTEGER PRIMARY KEY,
@@ -114,7 +153,8 @@ class DatabaseHelper {
         createdAt TEXT,
         name TEXT,
         likeCount INTEGER DEFAULT 0,
-        reportCount INTEGER DEFAULT 0 // 신고 수 필드 추가
+        reportCount INTEGER DEFAULT 0,
+        FOREIGN KEY (fk_memberNumber) REFERENCES $membersTable(memberNumber)
       )
     ''');
 
@@ -130,9 +170,45 @@ class DatabaseHelper {
     ''');
   }
 
-  // 초기 데이터 삽입
   static Future<void> _insertInitialData(Database db) async {
-    final List<Map<String, dynamic>> initialData = [
+    final List<Map<String, dynamic>> initialMembersData = [
+      {
+        'memberNumber': 1,
+        'password': 'password1',
+        'name': '회원1',
+        'phoneNumber': '010-1111-1111',
+        'registrationDate': DateTime.now().toIso8601String(),
+        'expirationDate':
+            DateTime.now().add(Duration(days: 365)).toIso8601String(),
+        'memberState': '정상'
+      },
+      {
+        'memberNumber': 2,
+        'password': 'password2',
+        'name': '회원2',
+        'phoneNumber': '010-2222-2222',
+        'registrationDate': DateTime.now().toIso8601String(),
+        'expirationDate':
+            DateTime.now().add(Duration(days: 365)).toIso8601String(),
+        'memberState': '정상'
+      },
+      {
+        'memberNumber': 3,
+        'password': 'password3',
+        'name': '회원3',
+        'phoneNumber': '010-3333-3333',
+        'registrationDate': DateTime.now().toIso8601String(),
+        'expirationDate':
+            DateTime.now().add(Duration(days: 365)).toIso8601String(),
+        'memberState': '정상'
+      },
+    ];
+
+    for (final data in initialMembersData) {
+      await db.insert(membersTable, data);
+    }
+
+    final List<Map<String, dynamic>> initialPostsData = [
       {
         'postID': 1,
         'fk_memberNumber': 1,
@@ -141,6 +217,8 @@ class DatabaseHelper {
         'content': '첫 번째 게시물 내용입니다.',
         'createdAt': DateTime.now().toIso8601String(),
         'name': '회원1',
+        'likeCount': 0,
+        'reportCount': 0
       },
       {
         'postID': 2,
@@ -150,6 +228,8 @@ class DatabaseHelper {
         'content': '두 번째 게시물 내용입니다.',
         'createdAt': DateTime.now().toIso8601String(),
         'name': '회원2',
+        'likeCount': 0,
+        'reportCount': 0
       },
       {
         'postID': 3,
@@ -159,17 +239,130 @@ class DatabaseHelper {
         'content': '세 번째 게시물 내용입니다.',
         'createdAt': DateTime.now().toIso8601String(),
         'name': '회원3',
+        'likeCount': 0,
+        'reportCount': 0
       },
     ];
 
-    for (final data in initialData) {
+    for (final data in initialPostsData) {
       await db.insert(postsTable, data);
     }
   }
 
+  // memberNumber로 특정 회원 정보 가져오기
+  Future<Map<String, dynamic>?> getMember(int memberNumber) async {
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(
+      membersTable,
+      where: 'memberNumber = ?',
+      whereArgs: [memberNumber],
+    );
+    if (maps.isNotEmpty) {
+      return maps.first;
+    } else {
+      return null;
+    }
+  }
+
+  static Future<void> insertMember(Member member) async {
+    final db = await database;
+    await db.insert(membersTable, member.toMap());
+  }
+
+  static Future<List<Member>> getMembers() async {
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(membersTable);
+    return List.generate(maps.length, (i) {
+      return Member(
+        memberNumber: maps[i]['memberNumber'],
+        password: maps[i]['password'],
+        name: maps[i]['name'],
+        phoneNumber: maps[i]['phoneNumber'],
+        registrationDate: DateTime.parse(maps[i]['registrationDate']),
+        expirationDate: DateTime.parse(maps[i]['expirationDate']),
+        memberState: maps[i]['memberState'],
+      );
+    });
+  }
+
+  static Future<int> getLastThreeDigits() async {
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(
+      membersTable,
+      columns: ['memberNumber'],
+      orderBy: 'memberNumber DESC',
+      limit: 1,
+    );
+    if (maps.isEmpty) {
+      return 0;
+    }
+    final lastMemberNumber = maps.first['memberNumber'] as int;
+    return lastMemberNumber % 1000;
+  }
+
+  static Future<void> deleteMember(int memberNumber) async {
+    final db = await database;
+    await db.delete(
+      membersTable,
+      where: 'memberNumber = ?',
+      whereArgs: [memberNumber],
+    );
+  }
+
+  static Future<void> updateMember(Member member) async {
+    final db = await database;
+    await db.update(
+      membersTable,
+      member.toMap(),
+      where: 'memberNumber = ?',
+      whereArgs: [member.memberNumber],
+    );
+  }
+
+  static Future<List<Member>> searchMembers(String searchQuery) async {
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(
+      membersTable,
+      where: 'name LIKE ? OR phoneNumber LIKE ?',
+      whereArgs: ['%$searchQuery%', '%$searchQuery%'],
+    );
+    return List.generate(maps.length, (i) {
+      return Member(
+        memberNumber: maps[i]['memberNumber'],
+        password: maps[i]['password'],
+        name: maps[i]['name'],
+        phoneNumber: maps[i]['phoneNumber'],
+        registrationDate: DateTime.parse(maps[i]['registrationDate']),
+        expirationDate: DateTime.parse(maps[i]['expirationDate']),
+        memberState: maps[i]['memberState'],
+      );
+    });
+  }
+
+  static Future<List<Member>> IDCheck(String id, String password) async {
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(
+      membersTable,
+      where: 'memberNumber = ? AND password = ?',
+      whereArgs: [id, password],
+    );
+    return List.generate(maps.length, (i) {
+      return Member(
+        memberNumber: maps[i]['memberNumber'],
+        password: maps[i]['password'],
+        name: maps[i]['name'],
+        phoneNumber: maps[i]['phoneNumber'],
+        registrationDate: DateTime.parse(maps[i]['registrationDate']),
+        expirationDate: DateTime.parse(maps[i]['expirationDate']),
+        memberState: maps[i]['memberState'],
+      );
+    });
+  }
+
   // 신고 수 업데이트 메서드 추가
   static Future<void> updateReportCount(int postID, int newReportCount) async {
-    await _database.update(
+    final db = await database;
+    await db.update(
       postsTable,
       {'reportCount': newReportCount},
       where: 'postID = ?',
@@ -179,7 +372,8 @@ class DatabaseHelper {
 
   // 특정 게시물의 신고 수 가져오기
   static Future<int> getReportCount(int postID) async {
-    final List<Map<String, dynamic>> maps = await _database.query(
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(
       postsTable,
       columns: ['reportCount'],
       where: 'postID = ?',
@@ -194,7 +388,8 @@ class DatabaseHelper {
 
   // 특정 유형의 게시물 가져오기
   static Future<List<Commu>> getPostsByType(String type) async {
-    final List<Map<String, dynamic>> maps = await _database.query(
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(
       postsTable,
       where: 'type = ?',
       whereArgs: [type],
@@ -224,17 +419,20 @@ class DatabaseHelper {
 
   // 게시물 삽입
   static Future<void> insertPost(Commu post) async {
-    await _database.insert(postsTable, post.toMap());
+    final db = await database;
+    await db.insert(postsTable, post.toMap());
   }
 
   // 댓글 삽입
   static Future<void> insertComment(Comment comment) async {
-    await _database.insert(commentsTable, comment.toMap());
+    final db = await database;
+    await db.insert(commentsTable, comment.toMap());
   }
 
   // 특정 게시물의 댓글 가져오기
   static Future<List<Comment>> getCommentsByPostID(int postID) async {
-    final List<Map<String, dynamic>> maps = await _database.query(
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(
       commentsTable,
       where: 'postID = ?',
       whereArgs: [postID],
@@ -252,7 +450,8 @@ class DatabaseHelper {
 
   // 모든 게시물 가져오기
   static Future<List<Commu>> getPosts() async {
-    final List<Map<String, dynamic>> maps = await _database.query(postsTable);
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(postsTable);
 
     List<Commu> posts = [];
     for (final map in maps) {
@@ -278,7 +477,8 @@ class DatabaseHelper {
 
   // 특정 게시물의 댓글 수 가져오기
   static Future<int> _getCommentCount(int postID) async {
-    final List<Map<String, dynamic>> maps = await _database.query(
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(
       commentsTable,
       where: 'postID = ?',
       whereArgs: [postID],
@@ -288,7 +488,8 @@ class DatabaseHelper {
 
   // 특정 게시물의 좋아요 수 가져오기
   static Future<int> getLikeCount(int postID) async {
-    final List<Map<String, dynamic>> maps = await _database.query(
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(
       postsTable,
       columns: ['likeCount'],
       where: 'postID = ?',
@@ -303,7 +504,8 @@ class DatabaseHelper {
 
   // 특정 게시물의 좋아요 수 업데이트
   static Future<void> updateLikeCount(int postID, int newLikeCount) async {
-    await _database.update(
+    final db = await database;
+    await db.update(
       postsTable,
       {'likeCount': newLikeCount},
       where: 'postID = ?',
@@ -313,13 +515,14 @@ class DatabaseHelper {
 
 // 게시물 삭제 메서드 추가
   static Future<void> deletePost(int postID) async {
-    await _database.delete(
+    final db = await database;
+    await db.delete(
       postsTable,
       where: 'postID = ?',
       whereArgs: [postID],
     );
     // 해당 게시물의 댓글도 삭제
-    await _database.delete(
+    await db.delete(
       commentsTable,
       where: 'postID = ?',
       whereArgs: [postID],
@@ -327,9 +530,9 @@ class DatabaseHelper {
   }
 
   // 게시물 검색
-  // 게시물 검색
   static Future<List<Commu>> searchPosts(String query) async {
-    final List<Map<String, dynamic>> maps = await _database.query(
+    final db = await database;
+    final List<Map<String, dynamic>> maps = await db.query(
       postsTable,
       where: 'title LIKE ? OR content LIKE ?', // 제목 또는 내용에 대한 조건
       whereArgs: ['%$query%', '%$query%'], // 쿼리를 제목과 내용에 대한 검색어로 대체
